+++ libft/ft_memmove.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memmove.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gvon-ah- <gvon-ah-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: gisrael <gisrael@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 16:03:55 by gvon-ah-          #+#    #+#             */
-/*   Updated: 2024/11/04 15:34:03 by gvon-ah-         ###   ########.fr       */
+/*   Updated: 2025/02/24 15:59:23 by gisrael          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (src == NULL && dst == NULL)
 		return (NULL);
 	if ((size_t)dst - (size_t)src < len)
 	{
-		i = len -1;
-		while (i >= 0 && i < len)
+		i = (int)len -1;
+		while (i >= 0 && i < (int)len)
 		{
 			(((unsigned char *)dst)[i] = ((unsigned char *)src)[i]);
 			i--;
@@ -30,7 +30,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	}
 	else
 	{
-		while (i < len)
+		while (i < (int)len)
 		{
 			(((unsigned char *)dst)[i] = ((unsigned char *)src)[i]);
 			i++;
